HBX-1610: Move class 'org.hibernate.cfg.reveng.ReverseEngineeringStrategyUtil' to 'org.hibernate.tool.internal.util.NameConverter'
